requirements.txt
uvicorn
fastapi
psycopg2-binary
flask
pandas
requests
python-dotenv
python-multipart
tqdm
cryptography